refactor: consolidate common styles into shared.styles.ts
- Create shared.styles.ts for reusable semantic patterns (section, link, bgPrimary, textPrimary)
- Refactor footer.styles.ts to re-export shared styles
- Refactor page.styles.ts to re-export shared section style
- Refactor projects.styles.ts to re-export shared link style
- Refactor 404 page to use shared styles instead of duplicating
- Reduce code duplication: 4 files now share common styles
- All 49 tests pass

This eliminates the last of the duplicated Bootstrap-style patterns.
